requirements.txt
tensorflow-cpu
fastapi
uvicorn
python-multipart
Pillow
numpy